requirements.txt
pytelegrambotapi
paramiko
pynacl
qrcode